refactor(db): adopt one-word module names
