+++ crates/db/src/collection/index.rs
