+++ crates/db/src/txn/registry/lenses.rs
